add markers + split code in folders
